Refactor ActivityPage: Move ActivityType enum to a separate file and enhance log fetching logic
