+++ app/(dashboard)/dashboard/activity/types.ts
@@ -1,0 +1,17 @@
+export enum ActivityType {
+    SIGN_UP = 'SIGN_UP',
+    SIGN_IN = 'SIGN_IN',
+    SIGN_OUT = 'SIGN_OUT',
+    UPDATE_PASSWORD = 'UPDATE_PASSWORD',
+    DELETE_ACCOUNT = 'DELETE_ACCOUNT',
+    UPDATE_ACCOUNT = 'UPDATE_ACCOUNT',
+    CREATE_TEAM = 'CREATE_TEAM',
+    REMOVE_TEAM_MEMBER = 'REMOVE_TEAM_MEMBER',
+    INVITE_TEAM_MEMBER = 'INVITE_TEAM_MEMBER',
+    ACCEPT_INVITATION = 'ACCEPT_INVITATION',
+  
+    DEPOSIT = 'DEPOSIT',
+    WITHDRAW = 'WITHDRAW',
+  
+    AUTOMATION_CHANGE = 'AUTOMATION_CHANGE',
+  }
